pyproject.toml
[project]
name = "RPI_BME280_sensor"
version = "0.1.2"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard-no-fastapi-cloud-cli]>=0.135.1",
    "pimoroni-bme280>=1.0.0",
    "smbus2>=0.6.0",
    "typer>=0.24.1",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    {include-group = "lint"}
]
lint = [
    "mypy==1.20.2",
	"ruff==0.15.12",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]

[tool.ruff.lint.pydocstyle]
convention = "google"
ignore-decorators = ["typing.overload"]

